Moved around serializer, llm, connections_manager into engine
